fix(hydration): drop over-cautious assertion after valid usecase found
dont throw an error if the prop was declared as a nested domain object
and instantiated as a domain object, but not one of the declared domain
object options.

we've found valid usecases which this blocks (e.g.,
DeclaredResourceReference passed into a
DeclaredGoogleAdsCampaignReference). This was an extra check that was
added to try and help users detect errors earlier. We've updated the
later check to convey the same information while still allowing the
other resolution techniques to operate instead.